[a5a-5e4] Restructured folders
