+++ Assets/Match-3/Script/Board/AddedTileInfo.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+public struct AddedTileInfo
+{
+    public Vector2Int position;
+    public int type;
+}
