+++ Assets/Match-3/Script/Board/MovedTileInfo.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+public class MovedTileInfo
+{
+    public Vector2Int from;
+    public Vector2Int to;
+}
