+++ Assets/Match-3/Script/Core.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0a301b236ec98f949853f576337d18ff
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
