+++ Assets/Match-3/Script/Settings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1ff798f7941fa1147adbd7f418c37b27
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
