+++ Assets/Match-3/Script/Settings/TilePrefabRepository.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+[CreateAssetMenu(fileName = "TilePrefabRepository", menuName = "Gameplay/TilePrefabRepository")]
+public class TilePrefabRepository : ScriptableObject
+{
+    public TileView[] tileTypePrefabList;
+}
