+++ Assets/Match-3/Script/Tile/Tile.cs
@@ -1,0 +1,6 @@
+﻿public class Tile
+{
+    public int id;
+    // TODO: Convert to enum
+    public int type;
+}
