+++ Assets/Match-3/Script/Tile/TileView.cs
@@ -1,0 +1,7 @@
+﻿using System;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class TileView : MonoBehaviour
+{
+}
