+++ Assets/Match-3/Settings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 34951ea9aa6346145b5200c9485178bd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
