--- Assets/Script/AddedTileInfo.cs
@@ -1,7 +1,0 @@
-﻿using UnityEngine;
-
-public struct AddedTileInfo
-{
-    public Vector2Int position;
-    public int type;
-}
