--- Assets/Script/MovedTileInfo.cs
@@ -1,7 +1,0 @@
-﻿using UnityEngine;
-
-public class MovedTileInfo
-{
-    public Vector2Int from;
-    public Vector2Int to;
-}
