--- Assets/Script/Tile.cs
@@ -1,6 +1,0 @@
-﻿public class Tile
-{
-    public int id;
-    // TODO: Convert to enum
-    public int type;
-}
