--- Assets/Script/TilePrefabRepository.cs
@@ -1,7 +1,0 @@
-﻿using UnityEngine;
-
-[CreateAssetMenu(fileName = "TilePrefabRepository", menuName = "Gameplay/TilePrefabRepository")]
-public class TilePrefabRepository : ScriptableObject
-{
-    public TileView[] tileTypePrefabList;
-}
