--- Assets/Script/TileView.cs
@@ -1,7 +1,0 @@
-﻿using System;
-using UnityEngine;
-using UnityEngine.UI;
-
-public class TileView : MonoBehaviour
-{
-}
